first draft of separate libraries
